Rewrite Turmoil serialization.
This is a series of small changes that grew into something a bit larger,
but I think cleaner. It removes Object.assign, for instance, which means
serialization and deserialization are far more explicit.

While making this change I also changed the way Turmoil instances are
constructed. If you're going to make a Turmoil from a Game, you can use
loadFromGame(), which  works great for first-time set-up. But when you
want to load from JSON, you no longer need a Game instance, which makes
sense, doesn't it? Why create a vestigial game just to create a new
instance? I hope that we can propagate this up to the Game and Player
instances, and get rid of the need for vestigial players just to, load a
Game from the database, or to clone one.

While at it, I made a backwards-compatible change to the save format
with the intent to remove the older ways.

This removes references to (most) non-serializable objects in Turmoil -
instead storing well-known indexes. It also makes a breaking change away
from a couple of typos in field names that can at least be removed now.

Testing was great at finding the bugs in my serialization, but I would
like to hear feedback on serialization test requests before you felt
comfortable making a change like this, especially because I would like
to propagate this to Player and Game.