Turmoil serialization updated. (#2038)
This is the first in a series of updates that will adopt the same
behavior: no direct access to constructors allowed, it will be private.
Deserialization will be renamed to, hey, `deserialize.`

This works just fine with `Turmoil`, `Dealer` and `Player` but
once you apply the same thing to `Game` you see just how tightly
coupled the current structure for serialization/deserialization
is tied to existing behavior. It won't be hard to fix, but it's clear
that the whole thing has to be done in a series of small PRs.
Hence, the easy one first.