Game may no longer have multiple players with the same player ID, and `firstPlayer` must be in the list of players.
This only happens in tests, and as you can see by this change, it's
*a lot* of tests, which casually use `[player, player]` as their
list of players in a game. This causes some super-weird test failures
when a player receives both ends of an action. We shouldn't be doing it
anyway.

There were surprisingly few tests that didn't like this: MiningGuild
and GameLoader. Both were easy enough to fix, ... no actually
MiningGuild was a pain in the butt. Mining Guild's issue was that
that it used maxOutOceans, which by default stops when it places
9 oceans -- but there are 12 ocean spaces! So I switched to a more
direct space filter.

GameLoader just needed dedicated Player IDs.