Add first Moon card, Mare Nectaris Mine.
This adds A LOT of dressing, but it's all tested. The one thing to note
is that MOON_CARD_MANIFEST is now part of ALL_MANIFESTS, which does
run the risk of introducing the cards into games, except the Dealer
isn't going to touch it since it's not registered.

But that's what I needed to do to get the card rendered in the UI.