Make SelectHowToPayForProjectCard work with Moon cards.
This commit does a few things: it deals with keeping track of what the
proper remaining steel and titanium a player may use and applies it
throughout the algorithm. That was the most complicated part -- no --
writing the tests was the most complicated part.

There's a bit of code in SHPTFPC which uses the player's played cards,
and includes the self replicating robots as its source for card data.
That just seems wrong. Of the many changes in this commit, that's the
one I'm most surprised is in the code, and the one I most want someone
to scrutinize and evaluate.

I rewrote the tests, but just a bit -- to take advantage of things like
Partial<>.

Finally, there's a server-side change which verifies that someone can't
overspend their steel, consequently, they can't go negative. Now, this
won't be possible anymore because of all the client-side changes, but
I have manually verified this with Luna Train Station.