Detangle id in player model.
See, here's an area where detangling the two models will help. Because
in PublicPlayerModel, id is always a color. (Might not be true at
game end, but duringmost of the game it is.)

But when used as the root Player model, id is in fact the PlayerId.

Wow, that took me quite some time to figure out.