Have input selection use playerView
Rather than rename all the fields in all the components at once, this
commit uses computed to ease the transition. Then I can do the change
for each one individually. But these have to be in a bunch because
there's a cross-reference between them thanks to the relationship
between PlayerInputFactory, AndOptions and OrOptions.