Completely rewrote alt venus server side handling to reuse Dry Deserts's code.
That's because I misread the rules; the resources gained by the track
don't have to be distinct.

Not as elegant, but it'll be much less code in the end.

The SelectDistinctResources action may still be useful later, because
the rule applied specifically to the Colony, which I have not yet
written.