Buff Mars Direct.
Mars Direct works because of all the Mars cards in the deck. When 8
cards are missing, that reduces the card's potential. A discussion was
raised to temporarily buff the card by 1.5MC per missing card.

Some Mars cards just won't get implemented, so this buff could in theory
be permanent in some form.