Add Pathfinders corporation Habitat Martae
I have thoughts about whether this should be included. I mean, it works,
but it means having to redefine how to count tags. It's kinda localized
to those cards that look at science tags, but is it worth having this
special-case? Should the new Player methods be applied everywhere?
Should there be a card.getTags() method that accomplishes this for you?