Render card data as a JSON file and use it to load cards.
This PR is optional based on whether 'experimental ui' is on. Once
people get a chance to try this without any issues and the flag is
removed, the front-end will no longer rely on ICard.

You can see this in action on terraforming-mars-kberg.herokuapp.com.
On that site you don't even need to set a flag, it's on by default.