Add type InputResponse to represent Array<Array<string>>
The goal is to eventually replace Array<Array<string>> JSON. Once this
happens, not only will responses be easier to understand, but also
simplify adding more information like and nested in or, or valuable
sidechannel information.

Even if the long goal isn't met (JSON as input format) the code is
clearer and slightly easier to understand.