Start moving input processing to the PlayerInputs.
I wanted to do this about a year ago, but this kind of change didn't
work because for some reason it created a dependency cycle that went
from server to client, back to server. Huh?

Now that client and server are mostly separate, this is safe to start
doing. \o/

Getting the processing out of a single method means making more room for
robust validation, and it takes some of the work out of `Player.ts`