Update translation key for "Declaration of Independence" description in German
The german description itself was not updated though. I'm not a german speaker, yet when comparing to other cards that use the "Requires that you have", the translation is the same. I guess it can stay as it is.

I did not find any other languages requiring this translation to be updated.