Fix broken undo.
The broken behavior* is due to the fact that when the server switches
between players it saves -- twice. Isn't that strange? It's how it
works. While that should be fixed, it's how it works.

If the app saved twice by saving from id 3 to 4, and then 4 to 5, that
would be great, but it's not what it was doing. It was saving v3, and
then saving as v3 again. Because save() also updates, this isn't
a real problem.

The problem is that both times, it increments `game.lastSaveId`, so during
the first save it saves v3 and bumps `game.lastSaveId` from 3 to 4, and during
the second save it saves v3 and bumps `game.lastSaveId` from 4 to 5.

This repair ensures that no matter what version the server expects to
save, it bumps `game.lastSaveId` to the number above it.

This is verified in the test with one of the last lines in the new test.
With old behavior, this test fails. I checked it.

What does all of this have to do with undo? Well, without trying to
describe it, basically, the database has versions 0,1,2,3, but
`game.lastSaveId` is 5, and then the server tries to load
`game.lastSaveId - 1`.

There you go.