Rewrote creating games for tests yet again.
I have mixed feelings about this approach. It introduces a fake form of
extension that doesn't really accomplish what I set out to do, which is
to provide an equivalent to TestPlayer. The difference is that
TestPlayers' parent, Player, has an accessible constructor, but Game
does not.

So, just not clear.