Remove most uses of setCustomGameOptions. (#4926)
setCustomGameOptions was pretty useful in early development, and made it
possible to test all the expansions without much extra work. At the
time, the structures for setting game options were hidden by private
members, and probably not a lot of Typescript savvy (e.g. the spread
operator, which helps. A LOT.)

The drawback to setCustomGameOptions is that it hides the actual set of
game options used in the test from the reader. It also adds expansions
that are not relevant to the tests.