Remove most uses of `tr` attributes by inferring it from behavior. (#4939)
* Move MoonSpaces to common.

They're not used in common, but this makes Behavior all based on just
Common.

* Remove most uses of `tr` attributes by inferring it from behavior.

Many uses of `tr` remain, and that's because in those cases, the
behavior does not describe the change. For example, some cards place 2
oceans, and Behavior is not set to handle that, hence, the repetition.

The algorithm is specifically defined as

If the card specifies tr, use that. Otherwise infer it from behavior. If
there's no behavior defined, then there's no TR computation.

* Greenery behavior attribute should include a TR bump