Add experimental language switcher
This moves the language selection from the preferences dialog to its
own spot on the sidebar. A downside is that this means it consumes
more vertical space, which is probably bad for mobile. A benefit,
though, is that it's immediately understandable, plus, when you select
a language the page just refreshes.

This reuses alot from LanguageSwitcher.