Replace array of arrays with JSON responses.
Having this type of structure will allow some features currently not
possible, including PlayerInputs nested more than one level deep, adding
a top-level interrupt-type response (think 'undo') and also adding a unique
request/response ID, which just adds a little bit of safety for
knowing that the response is genuinely for the PlayerInput that player
is waiting for.

This will improve comprehension at the cost of network bandwidth, but I
believe it is an acceptable tradeoff.

It's possible these validations are a little strict (so what if
additional fields are part of the input?) But we can loosen those.

Still pending are some new tests, and also some serious manual testing.