Add unplayed corporation cards to the discard pile.
This makes them available for Merger.

Also, this is moved to SelectInitialCards as it puts most of the initial
card setup in one place. There's probably a cleaner way to do this by
moving _all_ the selection out of SIC, but this is an easy way to test.

Partially fixes #5277