Update AncientShipyards.ts
Adjustment for ancient shipyards for solo mode only. The orginal effect is back and it is now a multiple "Indentured Workers" card, thus balanced as +8M€ for -1VP multiple times if you wish. So its fair for solo,

This would be the way how it should be copied over to the Darkside of the Moon Syndicate corp for solo mode only.