Update reserveUnits when a new card is selected (#5840)
Currently, the reserveUnits variable for playing a card is fixed to the first card selected and when the selected card is changed, the variable is not updated. This may allow a user to try and play a card they cannot afford or be unable to play a card that they can afford. The bug is fixed by updating reserveUnits as the selected card changes.