Remove Zod. (#6529)
Great idea, but it might have taken too much memory. It's not entirely
clear how much Zod took up, but it caused Heroku to stop compiling our
app. It also introduced a couple of errors that were going to be fun to
fix.

This breaks the red-rover bot, but that's also in development, and can
be addressed later.

Fixes #6528