Move most uses of card.production to card.productionBox.
This will facilitate better prediction for Cyberia Systems.
It has the side effect of making a better calculation for Specialized
Settlement and Robotic Workforce, in the case where the player gained
one energy production (e.g. on Terra Cimmeria.)

The final remnant of card.production is Small Open Pit Mine, which
lets a player choose between two options. This can't be reduced
to a single option, but as a single-case, it will probably be hard-coded
into dealing with Cyberia Systems.