First implementation of autopass
Autopass is only available when you are taking actions because of how autopass is stored.
It's connected with when the player takes actions, and is saved along with the rest of the
game. This means that if you are waiting for someone else to take actions, you can't change
your autopass setting.

If there was another database table that stored autopass and other settings, this could be
changed to be available at all times.