Milestone award image cleanup (#7307)
* Move all milestone and award images to one directory.

This is going to simplify a few things. First as you see, it simplifies
most of the declarations. After this, it's going to simplify slightly
resizing all the images so they have generally the same shape. All the
MA images if you look at them have extra content that isn't shown. That
will all be trimmed.

There are still more MAs that we can add, and this will just simplify
all of that.

* Trim all milestones and awards.

This makes it easier to create new milestones and awards since they fit exactly in the 140x83 grid.