Faster card lookups with new playedCards.
playedCards is no longer an array, because searching in an array is
O(n). It can be more efficient if this is a first-class object and
support for faster lookups.

There are still cases where the underlaying array is accessible; those
could eventually be filtered a bit, and maybe removed entirely, though
that's not vital. What would be useful though is to add more data to the
playedCards data structure. For instance, event cards have to be counted
all the time. Now that playedCards is controlled by data structure, that
can just be computed as cards are played (or removed.)

A lot of tests needed to change to support this. Sorry about that.
There's probably ways to reduce a bunch of this, but it's also not
vital.