Centralize stock and production code. Promote reuse.
stock and production are no longer hidden behind getters and setters,
and production is now writeable just as if it was readable.

This seems possibly a bad idea, but this is still a pretty small project
even after all these years, and that data does not need to be hidden
behind a linguistic control. And, it makes operations faster.