Milestones and Awards updates for Terra Cintra Novus based on the official expansion.
All awards were defined, just needed to be specified for this board.

For milestones, a few needed to be added.  For cases with small variations, defined alternatives within the same module as alternate exported symbols.  Introduced the notion of a TagBasedMilestone which many other milestones also fall under.  This helps eliminate defining new classes with no real behavioral/logic differences.